pyproject.toml
[project]
name = "qd-browser"
version = "0.1.0"
description = "基于 Playwright 的综合爬虫 CLI 工具"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "playwright>=1.50.0",
    "typer>=0.15.0",
    "beautifulsoup4>=4.12.0",
    "trafilatura>=1.13.0",
    "html2text>=2024.2.26",
    "rich>=13.9.0",
    "pydantic-settings>=2.8.0",
    "httpx>=0.28.0",
    "lxml>=5.3.0",
    "tldextract>=5.3.1",
    "requests>=2.33.1",
    "python-dotenv>=1.2.2",
]

[project.scripts]
qd-browser = "qd_browser.cli:app"

[tool.uv]
package = true

[dependency-groups]
dev = []
